Ensure efficient and atomic updates of kubeconfigs (#17)
* Ensure efficient and atomic updates of kubeconfigs

The various kubeconfig files are rewritten on every hook even if they
don't change. Additionally, due to the use of `kubectl` to write the
data using the correct fields, there is a mutli-second window in which
clients reading the file could see an empty or partial file. This
changes it to only update a given config file if it has actually changed
and to do so in an entirely atomic fashion which will not disrupt
clients which are actively reading the file.

Fixes [lp:1837930][]

[lp:1837930]: https://bugs.launchpad.net/charm-kubernetes-master/+bug/1837930

* Fix workflow

* Fix missing sudo

* Add some comments to the tests

* Blacken like a delicious salmon fillet